master/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator>=2.1.1
pydantic==2.6.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
httpx
boto3
pysendpulse>=0.1.8
